+++ examples/multiwindow.rs
@@ -14,7 +14,7 @@ fn main() {
         let context = glutin::ContextBuilder::new();
         let gl_window = glutin::GlWindow::new(window, context, &events_loop).unwrap();
         let _ = unsafe { gl_window.make_current() };
-        let gl = support::load(&gl_window);
+        let gl = support::load(&gl_window.context());
         let window_id = gl_window.id();
         windows.insert(window_id, (gl_window, gl));
     }