+++ src/api/android/mod.rs
@@ -24,7 +24,7 @@ mod ffi;
 
 struct AndroidContext {
     egl_context: EglContext,
-    stopped: Cell<bool>,
+    stopped: Option<Cell<bool>>,
 }
 
 pub struct Context(Arc<AndroidContext>);
@@ -39,7 +39,7 @@ impl android_glue::SyncEventHandler for AndroidSyncEventHandler {
             // native window surface already destroyed. EGL generates a BAD_SURFACE error in this situation.
             // Set stop to true to prevent swap_buffer call race conditions.
             android_glue::Event::TermWindow => {
-                self.0.stopped.set(true);
+                self.0.stopped.as_ref().unwrap().set(true);
             },
             _ => { return; }
         };
@@ -47,7 +47,7 @@ impl android_glue::SyncEventHandler for AndroidSyncEventHandler {
 }
 
 impl Context {
-    pub fn new(
+    pub unsafe fn new(
         window_builder: winit::WindowBuilder,
         events_loop: &winit::EventsLoop,
         pf_reqs: &PixelFormatRequirements,
@@ -56,7 +56,7 @@ impl Context {
     {
         let window = window_builder.build(events_loop)?;
         let gl_attr = gl_attr.clone().map_sharing(|c| &c.0.egl_context);
-        let native_window = unsafe { android_glue::get_native_window() };
+        let native_window = android_glue::get_native_window();
         if native_window.is_null() {
             return Err(OsError(format!("Android's native window is null")));
         }
@@ -66,7 +66,7 @@ impl Context {
             .and_then(|p| p.finish(native_window as *const _)));
         let ctx = Arc::new(AndroidContext {
             egl_context: context,
-            stopped: Cell::new(false),
+            stopped: Some(Cell::new(false)),
         });
 
         let handler = Box::new(AndroidSyncEventHandler(ctx.clone()));
@@ -74,20 +74,16 @@ impl Context {
         let context = Context(ctx.clone());
 
         events_loop.set_suspend_callback(Some(Box::new(move |suspended| {
-            ctx.stopped.set(suspended);
+            ctx.stopped.as_ref().unwrap().set(suspended);
             if suspended {
                 // Android has stopped the activity or sent it to background.
                 // Release the EGL surface and stop the animation loop.
-                unsafe {
-                    ctx.egl_context.on_surface_destroyed();
-                }
+                ctx.egl_context.on_surface_destroyed();
             } else {
                 // Android has started the activity or sent it to foreground.
                 // Restore the EGL surface and animation loop.
-                unsafe {
-                    let native_window = android_glue::get_native_window();
-                    ctx.egl_context.on_surface_created(native_window as *const _);
-                }
+                let native_window = android_glue::get_native_window();
+                ctx.egl_context.on_surface_created(native_window as *const _);
             }
         })));
 
@@ -95,11 +91,36 @@ impl Context {
     }
 
     #[inline]
+    pub unsafe fn new_context(
+        _el: &winit::EventsLoop,
+        pf_reqs: &PixelFormatRequirements,
+        gl_attr: &GlAttributes<&Context>,
+        _shareable_with_windowed_contexts: bool,
+    ) -> Result<Self, CreationError> {
+        let gl_attr = gl_attr.clone().map_sharing(|c| &c.0.egl_context);
+        let context = EglContext::new(
+            egl::ffi::egl::Egl,
+            pf_reqs,
+            &gl_attr,
+            egl::NativeDisplay::Android
+        )?;
+        let context = context.finish_pbuffer((1, 1))?;// TODO:
+        let ctx = Arc::new(AndroidContext {
+            egl_context: context,
+            stopped: None,
+        });
+        Ok(Context(ctx))
+    }
+
+    #[inline]
     pub unsafe fn make_current(&self) -> Result<(), ContextError> {
-        if !self.0.stopped.get() {
-            return self.0.egl_context.make_current();
+        if let Some(ref stopped) = self.0.stopped {
+            if stopped.get() {
+                return Err(ContextError::ContextLost);
+            }
         }
-        Err(ContextError::ContextLost)
+
+        self.0.egl_context.make_current()
     }
 
     #[inline]
@@ -118,10 +139,12 @@ impl Context {
 
     #[inline]
     pub fn swap_buffers(&self) -> Result<(), ContextError> {
-        if !self.0.stopped.get() {
-            return self.0.egl_context.swap_buffers();
+        if let Some(ref stopped) = self.0.stopped {
+            if stopped.get() {
+                return Err(ContextError::ContextLost);
+            }
         }
-        Err(ContextError::ContextLost)
+        self.0.egl_context.swap_buffers()
     }
 
     #[inline]
@@ -137,67 +160,5 @@ impl Context {
     #[inline]
     pub unsafe fn raw_handle(&self) -> egl::ffi::EGLContext {
         self.0.egl_context.raw_handle()
-    }
-}
-
-#[derive(Clone, Default)]
-pub struct PlatformSpecificHeadlessBuilderAttributes;
-
-pub struct HeadlessContext(EglContext);
-
-unsafe impl Send for HeadlessContext {}
-unsafe impl Sync for HeadlessContext {}
-
-impl HeadlessContext {
-    /// See the docs in the crate root file.
-    pub fn new(
-        dimensions: (u32, u32),
-        pf_reqs: &PixelFormatRequirements,
-        gl_attr: &GlAttributes<&HeadlessContext>,
-        _: &PlatformSpecificHeadlessBuilderAttributes,
-    ) -> Result<Self, CreationError>
-    {
-        let gl_attr = gl_attr.clone().map_sharing(|c| &c.0);
-        let context = EglContext::new(egl::ffi::egl::Egl,
-                                           pf_reqs,
-                                           &gl_attr,
-                                           egl::NativeDisplay::Android)?;
-        let context = context.finish_pbuffer(dimensions)?;     // TODO:
-        Ok(HeadlessContext(context))
-    }
-
-    #[inline]
-    pub unsafe fn make_current(&self) -> Result<(), ContextError> {
-        self.0.make_current()
-    }
-
-    #[inline]
-    pub fn is_current(&self) -> bool {
-        self.0.is_current()
-    }
-
-    #[inline]
-    pub fn get_proc_address(&self, addr: &str) -> *const () {
-        self.0.get_proc_address(addr)
-    }
-
-    #[inline]
-    pub fn swap_buffers(&self) -> Result<(), ContextError> {
-        self.0.swap_buffers()
-    }
-
-    #[inline]
-    pub fn get_api(&self) -> Api {
-        self.0.get_api()
-    }
-
-    #[inline]
-    pub fn get_pixel_format(&self) -> PixelFormat {
-        self.0.get_pixel_format()
-    }
-
-    #[inline]
-    pub unsafe fn raw_handle(&self) -> egl::ffi::EGLContext {
-        self.0.raw_handle()
     }
 }