+++ src/api/egl/mod.rs
@@ -20,6 +20,7 @@ use std::cell::Cell;
 pub mod ffi;
 
 /// Specifies the type of display passed as `native_display`.
+#[allow(dead_code)]
 pub enum NativeDisplay {
     /// `None` means `EGL_DEFAULT_DISPLAY`.
     X11(Option<ffi::EGLNativeDisplayType>),
@@ -42,6 +43,7 @@ pub struct Context {
     surface: Cell<ffi::egl::types::EGLSurface>,
     api: Api,
     pixel_format: PixelFormat,
+    #[cfg(target_os = "android")]
     config_id: ffi::egl::types::EGLConfig,
 }
 
@@ -169,10 +171,6 @@ impl Context {
         native_display: NativeDisplay,
     ) -> Result<ContextPrototype<'a>, CreationError>
     {
-        if opengl.sharing.is_some() {
-            unimplemented!()
-        }
-
         // calling `eglGetDisplay` or equivalent
         let display = get_native_display(&egl, native_display);
 
@@ -338,7 +336,7 @@ impl Context {
     // Restore the EGLContext.
     #[cfg(target_os = "android")]
     pub unsafe fn on_surface_created(&self, native_window: ffi::EGLNativeWindowType) {
-        if (self.surface.get() != ffi::egl::NO_SURFACE) {
+        if self.surface.get() != ffi::egl::NO_SURFACE {
             return;
         }
         self.surface.set(self.egl.CreateWindowSurface(self.display, self.config_id, native_window, ptr::null()));
@@ -357,7 +355,7 @@ impl Context {
     // The EGLContext is not destroyed so it can be restored later.
     #[cfg(target_os = "android")]
     pub unsafe fn on_surface_destroyed(&self) {
-        if (self.surface.get() == ffi::egl::NO_SURFACE) {
+        if self.surface.get() == ffi::egl::NO_SURFACE {
             return;
         }
         let ret = self.egl.MakeCurrent(self.display, ffi::egl::NO_SURFACE, ffi::egl::NO_SURFACE, ffi::egl::NO_CONTEXT);
@@ -444,22 +442,27 @@ impl<'a> ContextPrototype<'a> {
     fn finish_impl(self, surface: ffi::egl::types::EGLSurface)
                    -> Result<Context, CreationError>
     {
+        let share = match self.opengl.sharing {
+            Some(ctxt) => ctxt.context,
+            None => ptr::null()
+        };
+
         let context = unsafe {
             if let Some(version) = self.version {
                 create_context(&self.egl, self.display, &self.egl_version,
                                     &self.extensions, self.api, version, self.config_id,
-                                    self.opengl.debug, self.opengl.robustness)?
+                                    self.opengl.debug, self.opengl.robustness, share)?
 
             } else if self.api == Api::OpenGlEs {
                 if let Ok(ctxt) = create_context(&self.egl, self.display, &self.egl_version,
                                                  &self.extensions, self.api, (2, 0), self.config_id,
-                                                 self.opengl.debug, self.opengl.robustness)
+                                                 self.opengl.debug, self.opengl.robustness, share)
                 {
                     ctxt
                 } else if let Ok(ctxt) = create_context(&self.egl, self.display, &self.egl_version,
                                                         &self.extensions, self.api, (1, 0),
                                                         self.config_id, self.opengl.debug,
-                                                        self.opengl.robustness)
+                                                        self.opengl.robustness, share)
                 {
                     ctxt
                 } else {
@@ -469,19 +472,19 @@ impl<'a> ContextPrototype<'a> {
             } else {
                 if let Ok(ctxt) = create_context(&self.egl, self.display, &self.egl_version,
                                                  &self.extensions, self.api, (3, 2), self.config_id,
-                                                 self.opengl.debug, self.opengl.robustness)
+                                                 self.opengl.debug, self.opengl.robustness, share)
                 {
                     ctxt
                 } else if let Ok(ctxt) = create_context(&self.egl, self.display, &self.egl_version,
                                                         &self.extensions, self.api, (3, 1),
                                                         self.config_id, self.opengl.debug,
-                                                        self.opengl.robustness)
+                                                        self.opengl.robustness, share)
                 {
                     ctxt
                 } else if let Ok(ctxt) = create_context(&self.egl, self.display, &self.egl_version,
                                                         &self.extensions, self.api, (1, 0),
                                                         self.config_id, self.opengl.debug,
-                                                        self.opengl.robustness)
+                                                        self.opengl.robustness, share)
                 {
                     ctxt
                 } else {
@@ -497,7 +500,8 @@ impl<'a> ContextPrototype<'a> {
             surface: Cell::new(surface),
             api: self.api,
             pixel_format: self.pixel_format,
-            config_id: self.config_id
+            #[cfg(target_os = "android")]
+            config_id: self.config_id,
         })
     }
 }
@@ -664,7 +668,7 @@ unsafe fn create_context(egl: &ffi::egl::Egl, display: ffi::egl::types::EGLDispl
                          egl_version: &(ffi::egl::types::EGLint, ffi::egl::types::EGLint),
                          extensions: &[String], api: Api, version: (u8, u8),
                          config_id: ffi::egl::types::EGLConfig, gl_debug: bool,
-                         gl_robustness: Robustness)
+                         gl_robustness: Robustness, share: ffi::EGLContext)
                          -> Result<ffi::egl::types::EGLContext, CreationError>
 {
     let mut context_attributes = Vec::with_capacity(10);
@@ -771,7 +775,7 @@ unsafe fn create_context(egl: &ffi::egl::Egl, display: ffi::egl::types::EGLDispl
 
     context_attributes.push(ffi::egl::NONE as i32);
 
-    let context = egl.CreateContext(display, config_id, ptr::null(),
+    let context = egl.CreateContext(display, config_id, share,
                                     context_attributes.as_ptr());
 
     if context.is_null() {