+++ src/platform/linux/wayland.rs
@@ -13,7 +13,7 @@ pub struct Context {
 }
 
 impl Context {
-    pub fn new(
+    pub unsafe fn new(
         window_builder: winit::WindowBuilder,
         events_loop: &winit::EventsLoop,
         pf_reqs: &PixelFormatRequirements,
@@ -23,18 +23,22 @@ impl Context {
         let window = window_builder.build(events_loop)?;
         let logical_size = window.get_inner_size().unwrap();
         let (w, h) = (logical_size.width, logical_size.height);
-        let surface = window.get_wayland_surface().unwrap();
-        let egl_surface = unsafe { wegl::WlEglSurface::new_from_raw(surface as *mut _, w as i32, h as i32) };
+        let surface = window.get_wayland_surface();
+        let surface = match surface {
+            Some(s) => s,
+            None => return Err(CreationError::NotSupported("Wayland not found")),
+        };
+        let egl_surface = wegl::WlEglSurface::new_from_raw(surface as *mut _, w as i32, h as i32);
         let context = {
-            let libegl = unsafe { dlopen::dlopen(b"libEGL.so\0".as_ptr() as *const _, dlopen::RTLD_NOW) };
+            let libegl = dlopen::dlopen(b"libEGL.so\0".as_ptr() as *const _, dlopen::RTLD_NOW);
             if libegl.is_null() {
                 return Err(CreationError::NotSupported("could not find libEGL"));
             }
             let egl = ::api::egl::ffi::egl::Egl::load_with(|sym| {
                 let sym = CString::new(sym).unwrap();
-                unsafe { dlopen::dlsym(libegl, sym.as_ptr()) }
+                dlopen::dlsym(libegl, sym.as_ptr())
             });
-            let gl_attr = gl_attr.clone().map_sharing(|_| unimplemented!()); // TODO
+            let gl_attr = gl_attr.clone().map_sharing(|c| &c.context);
             let native_display = egl::NativeDisplay::Wayland(Some(
                 window.get_wayland_display().unwrap() as *const _
             ));
@@ -48,6 +52,7 @@ impl Context {
         Ok((window, context))
     }
 
+    #[inline]
     pub fn resize(&self, width: u32, height: u32) {
         self.egl_surface.resize(width as i32, height as i32, 0, 0);
     }